third_party/blink/web_tests/external/wpt/css/css-scoping/has-slotted-functional-changing-002.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Changing content targetting :has-slotted(...) through a single shadow root
  assert_equals: expected "rgb(0, 0,  255)" but got "rgb(0, 255, 0)"
Harness: the test ran to completion.
